IoT_Project/requirements.txt
paho-mqtt
cherrypy
requests
python-telegram-bot
matplotlib
csv